+++ app/src/main/java/de/passwordvault/view/fragments/EntriesFragment.java
@@ -22,7 +22,6 @@ import android.widget.ListView;
 import android.widget.PopupMenu;
 import android.widget.Toast;
 import java.util.ArrayList;
-
 import de.passwordvault.R;
 import de.passwordvault.model.Observable;
 import de.passwordvault.model.Observer;